% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-survmat.R
\name{bind_survmat}
\alias{bind_survmat}
\title{Row-bind survmat objects on a common time grid}
\usage{
bind_survmat(...)
}
\arguments{
\item{...}{One or more \code{survmat} objects.}
}
\value{
A combined \code{survmat}.
}
\description{
Row-bind survmat objects on a common time grid
}
