% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_tvrmst_diff_reps}
\alias{bootstrap_tvrmst_diff_reps}
\title{Summarize Delta_c(s, tau) from a list of replicate survival matrices}
\usage{
bootstrap_tvrmst_diff_reps(reps, time, s_grid, tau, eps = 1e-08, conf = 0.95)
}
\arguments{
\item{reps}{List of length R, each element list(S1=..., S0=...)}

\item{time}{Time grid}

\item{s_grid}{Landmark grid}

\item{tau}{Window length}

\item{eps}{Stability threshold}

\item{conf}{Confidence level}
}
\value{
data.frame(s, estimate, lower, upper)
}
\description{
Summarize Delta_c(s, tau) from a list of replicate survival matrices
}
\examples{
set.seed(1)
t <- seq(0, 5, by = 1)
s_grid <- c(0, 1, 2)
reps <- lapply(1:10, function(i) {
  list(
    S1 = matrix(exp(-(0.15 + runif(1, -0.02, 0.02)) * t), nrow = 1),
    S0 = matrix(exp(-(0.20 + runif(1, -0.02, 0.02)) * t), nrow = 1)
  )
})
bootstrap_tvrmst_diff_reps(reps, t, s_grid, tau = 2, conf = 0.9)
}
