% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_delta_curve}
\alias{plot_delta_curve}
\title{Plot a delta curve}
\usage{
plot_delta_curve(
  grid,
  delta,
  title = "Delta curve",
  xlab = "t",
  ylab = "Delta"
)
}
\arguments{
\item{grid}{X-axis grid.}

\item{delta}{Y values.}

\item{title}{Plot title.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a delta curve
}
