% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rmst_curve}
\alias{plot_rmst_curve}
\title{Plot RMST curves over tau for multiple series}
\usage{
plot_rmst_curve(rmst_curve_df, title = "RMST curve")
}
\arguments{
\item{rmst_curve_df}{Output of \code{rmst_curve()}.}

\item{title}{Plot title.}
}
\description{
Plot RMST curves over tau for multiple series
}
\examples{
t <- seq(0, 5, by = 1)
S <- rbind(A = exp(-0.2 * t), B = exp(-0.3 * t))
rc <- rmst_curve(S, t)
if (requireNamespace("ggplot2", quietly = TRUE)) {
  plot_rmst_curve(rc)
}
}
