% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvrmst-package.R
\docType{package}
\name{tvrmst-package}
\alias{tvrmst-package}
\title{tvrmst: Time-Varying RMST from Survival Matrices}
\description{
Utilities for RMST and time-varying RMST quantities computed from survival curves provided on a time grid. Model-agnostic accepts only time vector and survival matrices, returning RMST-based quantities and bootstrap summaries.
}
\author{
\strong{Maintainer}: Imad EL BADISY \email{elbadisyimad@gmail.com}

}
\keyword{internal}
