% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvrmst.R
\name{tvrmst_cond}
\alias{tvrmst_cond}
\title{Conditional tvRMST}
\usage{
tvrmst_cond(
  S,
  time,
  s_grid,
  tau,
  eps = 1e-08,
  statistic = c("mean", "median", "unit")
)
}
\arguments{
\item{S}{Survival matrix/data.frame (rows = units, cols = time).}

\item{time}{Numeric strictly increasing time vector.}

\item{s_grid}{Landmark times.}

\item{tau}{Positive window length.}

\item{eps}{Positivity threshold for \code{S(s)}.}

\item{statistic}{One of \code{"mean"}, \code{"median"}, \code{"unit"}.}
}
\value{
For summary statistics: data.frame \code{s}, \code{estimate}.
For \code{statistic = "unit"}: numeric matrix n_units x length(s_grid).
}
\description{
Conditional tvRMST
}
