% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-survmat.R
\name{bind_survmat}
\alias{bind_survmat}
\title{Row-bind survmat objects on a common time grid}
\usage{
bind_survmat(...)
}
\arguments{
\item{...}{One or more \code{survmat} objects.}
}
\value{
A combined \code{survmat}.
}
\description{
Row-bind survmat objects on a common time grid
}
\examples{
time <- c(0, 1, 2)
xA <- as_survmat(matrix(c(1.0, 0.9, 0.7), nrow = 1), time, group = "A")
xB <- as_survmat(matrix(c(1.0, 0.8, 0.6), nrow = 1), time, group = "B")
x <- bind_survmat(xA, xB)
nobs_survmat(x)
}
