% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_tvrmst_diff_reps}
\alias{bootstrap_tvrmst_diff_reps}
\title{Summarize Delta_c(s, tau) from a list of replicate survival matrices}
\usage{
bootstrap_tvrmst_diff_reps(t, reps, s_grid, tau, eps = 0.05, conf = 0.95)
}
\arguments{
\item{t}{Time grid}

\item{reps}{List of length R, each element list(S1=..., S0=...)}

\item{s_grid}{Landmark grid}

\item{tau}{Window length}

\item{eps}{Stability threshold}

\item{conf}{Confidence level}
}
\value{
data.frame(s, estimate, lower, upper)
}
\description{
Summarize Delta_c(s, tau) from a list of replicate survival matrices
}
