% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{check_identities}
\alias{check_identities}
\title{Check RMST identity}
\usage{
check_identities(
  S,
  time,
  s_grid,
  tau,
  eps = 1e-08,
  statistic = c("mean", "unit")
)
}
\arguments{
\item{S}{Survival matrix/data.frame (rows = units, cols = time).}

\item{time}{Numeric strictly increasing time vector.}

\item{s_grid}{Landmark times.}

\item{tau}{Positive window length.}

\item{eps}{Positivity threshold.}

\item{statistic}{One of \code{"mean"} or \code{"unit"}.}
}
\value{
\code{statistic = "mean"}: data.frame with \code{s}, \code{max_abs_error}.
\code{statistic = "unit"}: data.frame with \code{s}, \code{unit}, \code{max_abs_error}.
}
\description{
Checks \code{RMST(s+tau)-RMST(s) = S(s) * mu_c(s,tau)}.
}
