% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-survmat.R
\name{nobs_survmat}
\alias{nobs_survmat}
\title{Number of rows in a survmat}
\usage{
nobs_survmat(x)
}
\arguments{
\item{x}{A \code{survmat} object.}
}
\value{
Number of subjects.
}
\description{
Number of rows in a survmat
}
\examples{
time <- c(0, 1, 2)
S <- rbind(
  c(1.0, 0.8, 0.6),
  c(1.0, 0.7, 0.5)
)
x <- as_survmat(S, time)
nobs_survmat(x)
}
