% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_delta_curve}
\alias{plot_delta_curve}
\title{Plot a delta curve}
\usage{
plot_delta_curve(
  grid,
  delta,
  title = "Delta curve",
  xlab = "t",
  ylab = "Delta",
  x_scale = 1,
  y_scale = 1,
  x_unit = NULL,
  y_unit = NULL
)
}
\arguments{
\item{grid}{X-axis grid.}

\item{delta}{Y values.}

\item{title}{Plot title.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{x_scale}{Positive divisor applied to the x-axis values before plotting.}

\item{y_scale}{Positive divisor applied to the y-axis values before plotting.}

\item{x_unit}{Optional unit label appended to the x-axis title.}

\item{y_unit}{Optional unit label appended to the y-axis title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a delta curve
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) withAutoprint(\{ # examplesIf}
time <- c(0, 1, 2)
xA <- as_survmat(rbind(c(1.0, 0.9, 0.7), c(1.0, 0.8, 0.6)), time, group = c("A", "A"))
xB <- as_survmat(rbind(c(1.0, 0.95, 0.8), c(1.0, 0.85, 0.7)), time, group = c("B", "B"))
d <- rmst_delta(xA, xB)
plot_delta_curve(d$time, d$delta)
\dontshow{\}) # examplesIf}
}
