% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rmst_curve}
\alias{plot_rmst_curve}
\title{Plot RMST curves over tau for multiple series}
\usage{
plot_rmst_curve(rmst_curve_df, title = "RMST curve")
}
\arguments{
\item{rmst_curve_df}{Output of \code{rmst_curve()}.}

\item{title}{Plot title.}
}
\description{
Plot RMST curves over tau for multiple series
}
