% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rmst_delta}
\alias{plot_rmst_delta}
\title{Plot Delta RMST curve over tau}
\usage{
plot_rmst_delta(rmst_delta_df, title = "Delta RMST curve")
}
\arguments{
\item{rmst_delta_df}{Output of \code{rmst_delta_curve()}.}

\item{title}{Plot title.}
}
\description{
Plot Delta RMST curve over tau
}
