% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rmst_individual_by_group}
\alias{plot_rmst_individual_by_group}
\title{Plot individual dynamic RMST curves by group}
\usage{
plot_rmst_individual_by_group(
  res,
  group,
  n_show_per_group = 30,
  title = "Individual dynamic RMST by group"
)
}
\arguments{
\item{res}{Result from \code{\link[=rmst_dynamic]{rmst_dynamic()}} containing \code{individual} and \code{time}.}

\item{group}{Group vector aligned with rows of \code{res$individual}.}

\item{n_show_per_group}{Maximum number of individual curves shown per group.}

\item{title}{Plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot individual dynamic RMST curves by group
}
