% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rmst_mean}
\alias{plot_rmst_mean}
\title{Plot mean dynamic RMST curves by group}
\usage{
plot_rmst_mean(rmst_dynamic_df, group, title = "Mean RMST by group")
}
\arguments{
\item{rmst_dynamic_df}{Output of \code{rmst_dynamic()}.}

\item{group}{Group labels (length = number of series).}

\item{title}{Plot title.}
}
\description{
Plot mean dynamic RMST curves by group
}
\examples{
t <- seq(0, 5, by = 1)
S <- cbind(A = exp(-0.2 * t), B = exp(-0.3 * t), C = exp(-0.25 * t))
rd <- rmst_dynamic(t, S)
if (requireNamespace("ggplot2", quietly = TRUE)) {
  plot_rmst_mean(rd, group = c("G1", "G2", "G1"))
}
}
