% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rmst_two_arms}
\alias{plot_rmst_two_arms}
\title{Plot mean RMST curves for two arms}
\usage{
plot_rmst_two_arms(
  xA,
  xB,
  labels = c("Arm A", "Arm B"),
  title = "Dynamic RMST (Two Arms)"
)
}
\arguments{
\item{xA}{A \code{survmat} object for arm A.}

\item{xB}{A \code{survmat} object for arm B.}

\item{labels}{Two legend labels.}

\item{title}{Plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot mean RMST curves for two arms
}
