% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_survival_curves}
\alias{plot_survival_curves}
\title{Plot two survival curves (first column only) as ggplot}
\usage{
plot_survival_curves(
  t,
  S0,
  S1,
  labels = c("Control", "Treatment"),
  title = "Survival curves"
)
}
\arguments{
\item{t}{Time grid.}

\item{S0}{Survival matrix for control arm.}

\item{S1}{Survival matrix for treatment arm.}

\item{labels}{Character vector of length 2 for legend labels.}

\item{title}{Plot title.}
}
\description{
Plot two survival curves (first column only) as ggplot
}
