% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_tvrmst}
\alias{plot_tvrmst}
\title{Plot conditional tvRMST curves for two arms (first series by default)}
\usage{
plot_tvrmst(
  mu0_df,
  mu1_df,
  labels = c("Control", "Treatment"),
  series_col = NULL,
  title = "Conditional tvRMST mu_c(s,tau) by arm"
)
}
\arguments{
\item{mu0_df}{Output of \code{tvrmst_cond()} for arm 0.}

\item{mu1_df}{Output of \code{tvrmst_cond()} for arm 1.}

\item{labels}{Character vector of length 2 for legend labels.}

\item{series_col}{Optional column name to plot.}

\item{title}{Plot title.}
}
\description{
Plot conditional tvRMST curves for two arms (first series by default)
}
