% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_tvrmst_diff}
\alias{plot_tvrmst_diff}
\title{Plot Delta_c(s,tau) with optional CI (estimate/lower/upper)}
\usage{
plot_tvrmst_diff(delta_df, title = "Delta_c(s,tau)")
}
\arguments{
\item{delta_df}{Output of \code{tvrmst_diff()} or a data.frame with columns
\code{s}, \code{estimate}, \code{lower}, \code{upper}.}

\item{title}{Plot title.}
}
\description{
Plot Delta_c(s,tau) with optional CI (estimate/lower/upper)
}
