% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst_dynamic}
\alias{rmst_dynamic}
\title{Dynamic RMST on the provided time grid}
\usage{
rmst_dynamic(S, time)
}
\arguments{
\item{S}{Survival predictions as matrix/data.frame with rows = units and
columns = time points.}

\item{time}{Numeric strictly increasing time vector.}
}
\value{
Numeric matrix of shape n_units x length(time).
}
\description{
Dynamic RMST on the provided time grid
}
