% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvrmst.R
\name{rmst_window}
\alias{rmst_window}
\title{Window RMST difference W(s, tau) = \eqn{\int_s^{s+\tau} (S_1 - S_0)\,du}}
\usage{
rmst_window(t, S1, S0, s_grid, tau)
}
\arguments{
\item{t}{Time grid.}

\item{S1}{Survival matrix (n_time x n_series) for arm 1.}

\item{S0}{Survival matrix (n_time x n_series) for arm 0.}

\item{s_grid}{Nonnegative landmark grid.}

\item{tau}{Positive window length.}
}
\value{
data.frame(s, <series...>)
}
\description{
Window RMST difference W(s, tau) = \eqn{\int_s^{s+\tau} (S_1 - S_0)\,du}
}
