% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvrmst.R
\name{tvrmst_diff}
\alias{tvrmst_diff}
\title{Difference in conditional tvRMST between arms}
\usage{
tvrmst_diff(
  S1,
  S0,
  time,
  s_grid,
  tau,
  eps = 1e-08,
  statistic = c("mean", "median", "unit")
)
}
\arguments{
\item{S1}{Arm-1 survival matrix/data.frame (rows = units, cols = time).}

\item{S0}{Arm-0 survival matrix/data.frame (rows = units, cols = time).}

\item{time}{Numeric strictly increasing time vector.}

\item{s_grid}{Landmark times.}

\item{tau}{Positive window length.}

\item{eps}{Positivity threshold.}

\item{statistic}{One of \code{"mean"}, \code{"median"}, \code{"unit"}.}
}
\value{
For summary statistics: data.frame \code{s}, \code{estimate}.
For \code{statistic = "unit"}: numeric matrix n_units x length(s_grid).
}
\description{
Difference in conditional tvRMST between arms
}
