% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvrmst.R
\name{tvrmst_diff}
\alias{tvrmst_diff}
\title{Conditional tvRMST difference Delta_c(s, tau) = mu_c1(s, tau) - mu_c0(s, tau)}
\usage{
tvrmst_diff(t, S1, S0, s_grid, tau, eps = 0.05)
}
\arguments{
\item{t}{Time grid.}

\item{S1}{Survival matrix for arm 1.}

\item{S0}{Survival matrix for arm 0.}

\item{s_grid}{Landmark grid.}

\item{tau}{Window length.}

\item{eps}{Stability threshold.}
}
\value{
data.frame(s, delta_c) if single series, else data.frame(s, <series...>)
}
\description{
Conditional tvRMST difference Delta_c(s, tau) = mu_c1(s, tau) - mu_c0(s, tau)
}
